Checkout V2: Expose AVS and CVV results for purchases
This adds AVS and CVV result details to `purchase`. Since the gateway
only returns AVS and CVV information when performing the initial
authorization, the responses from both `authorize` and `capture` are
now merged in `purchase`.

Also, this fixes a failing test by correcting how the adapter sends
address1 and address2.

Closes #2619