EBANX: Pass person_type and name for stored cards
Also conditions person_responsible element more correctly, and allows
the customer name passed to use the name associated with the billing
address when it is not available from the payment method, such as when
the payment method is a string for a stored card.

Four negative test cases are failing due to changes to test card numbers
we have yet to receive an update for.

Unit:
15 tests, 51 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote:
20 tests, 52 assertions, 4 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
80% passed