Safe Charge: Provisions adapter for 3DS purchases
Safe Charge has implemented a system by which they determine whether
purchases should be routed through the 3DS flow or not. If the merchant
is not approved for 3DS, transactions will occur as before, outside of
the 3DS flow. If a merchant is approved for 3DS, Sage Charge will
evaluate the card for enrollment status. If the card is not enrolled, it
will be processed as a 'regular' transaction. If the card is enrolled,
the Safe Charge response will include the 'xid', 'ascurl', and the
'pareq'.

Remote Tests:
21 tests, 62 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
100% passed

Unit Tests:
17 tests, 76 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
100% passed