Prevent frozen objects being assigned as wiredump device
The wiredump device takes an object that quacks like an IO stream (e.g. File or String) and passes it along to Net:HTTP#set_debug_output where it is used by the private `D` method for appending log messages, as seen here: https://github.com/ruby/ruby/blob/c91cb76f8d84b2963f6ede2ef445ad46a6104216/lib/net/http.rb#L1567

If the device is set to an empty string literal (i.e. "") and the frozen string literal magic comment is added to the file, this causes `RuntimeError: can't modify frozen String` exceptions. The stacktrace can be confusing since setting up the gateway object and using it usually happens in different places in the application code.

This PR aims to fail early if such a situation occurs and provide a stacktrace that clearly points to the cause of the error.