PayU Latam: Require payment_country on initialize
Since merchant accounts must be tied to a payment country, this finally
makes payment_country required for initalizing the gateway, and removes
support for overriding it with an option field.

Also, tests for refund and void are renamed and expect failure as they
are not currently supported in the sandbox. When they become supported,
these should fail, flagging them to be updated as passing tests.

Declined and pending test cards are now throwing a new unexpected error,
unrelated to these changes.

Closes #2663

Remote:
18 tests, 47 assertions, 3 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
83.3333% passed

Unit:
22 tests, 83 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed