FirstData E4: Override ECI value for Apple Pay transactions with Discover
Discover has a hard requirement that the ECI value for all Apple Pay
transactions be set to exactly 4, regardless of what is or is not
provided in the Apple PKPaymentToken. This is under subject of merchant
fees and penalties if not, according to people I've talked to at
Discover.

The Discover specification reads:

https://www.dropbox.com/s/fnmofdfnsc37wuj/Screenshot%202017-11-21%2017.43.16.png?dl=0

Yes, that's mostly meaningless, but "field 61 position 9" is what
gateways map the incoming ECI value to. I've confirmed this information
with contacts at FirstData, which said basically "yeah, this is a
Discover requirement, and you must send it in differently, we won't do
it for you."