Stripe: Fix Partial Application Fee Refunds
Previously, when a partial application fee refund was requested, the
connected stripe_account option was sent in the authentication header
for both the fee retrieval call AND the application fee refund itself.
Since the fee exists only on the main Platform account, sending the
connected stripe_account for the refund meant it was looking in the
wrong place, and failed. Now we remove the stripe_account field before
the partial refund, so it is not sent in headers.

This also switches to a method of fetching the transaction to obtain the
id of the application fee, rather than querying all application fees
associated with the transaction, since it simplifies things, and the
prior method did not seem to have any benefits over this (having
multiple application fees doesn't seem possible currently). Error
messaging around this has been updated as well.

This also involved updating the aging remote_stripe_connect_tests to use
the stripe_destination in fixtures as the stripe_account option, and to
remove assertions that are now invalid due to the changes to response
structure in https://stripe.com/docs/upgrades#2013-08-13
Unit tests were likewise updated for the partial fee refund method
changes.

Remote (two failing likely due to unrelated stale test data):
59 tests, 253 assertions, 2 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
96.6102% passed

Remote Connect:
3 tests, 10 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Unit:
122 tests, 651 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed