Set the `receipt_email` attribute of charges to `options[:email]`
I feel like this is how this should have always worked. The `:email`
option is a core, standardize option across Active Merchant, and is the
"customer's email address", meanwhile, the `receipt_email` attribute on
a Stripe charge is the "address a customer's receipt was sent to."
Definitely sounds like the same thing to me.

I've left the existing, one-off option `receipt_email` be, and it will
take precedence over the `email` option, which should maintain
backwards compatibility.