Ogone: add tests for scrubbing
Five remote tests fail, two involving extended options for purchases,
two on storing, and one on credits. None of these are impacted by this
patch series, and fixing them is beyond the scope of this series. (The
errors appears to be some sort of transient issue with their test card
processing.)

Unit:
48 tests, 206 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote:
27 tests, 91 assertions, 5 failures, 1 errors, 0 pendings, 0 omissions, 0 notifications
77.7778% passed