Add support for Ezidebit gateway
Ezidebit only seems to implement `purchase` and `store` via the SOAP API and not the usual suspects of `authorise`, `capture`, `void` and `refund`.

When using `purchase` we return the `receipt_id` and the `exchange_payment_id` returned values separated by a `|` (pipe) character so that a customer can have access to both and select which one they want to use in their downstream applications.