CenPos: Switch to new endpoint URL
In addition to changing the endpoint, something in Active Merchant
or this adapter isn't dealing well with Gzip'd responses, so I've
removed that header for now. (Requests at the old endpoint weren't ever
Gzip'd, so the old path was effectively untested.) I also patched one
remote test that was trivial to fix while I was in here.

The six remote test failures are identical with or without this patch.
(Well, you'll get seven without it due to the above-mentioned trivial
fix.)

Unit:
23 tests, 98 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote:
23 tests, 51 assertions, 6 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
73.913% passed