Payeezy: Update transaction method when using stored cards
For reference based transactions that accept stored cards, we'll now
pass `credit_card` instead of `token` in the `method` option. It's
not entirely clear from Payeezy's documentation that this is the
correct value to use, but with `credit_card` we no longer see the "Token
information is missing" error message on `capture`, `void`, and `refund`
calls.

Closes #2770

Unit:
32 tests, 153 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote:
30 tests, 118 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed