Adyen: Support store action
As part of implementation, I preserved the behavior of retaining
original psp references for refunds, and account for backwards
compatability with pre-existing non-compound authorizations. Also
removes an irrelevant test case (a capture should never have a compound
authorization) and switches Verify to a 0 amount, since it is supported.

Remote:
32 tests, 81 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Unit:
18 tests, 86 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed