CardStream: Change `refund` to use `REFUND_SALE`
Per CardStream, specifying `REFUND` as the action performs a general /
independent credit against the card which may result in chargebacks. Instead,
the `REFUND_SALE` action can be used refund a `SALE` transaction that has settled.
When a `SALE` transaction has not settled, a `CANCEL` request must be made that
will refund the original `SALE`.

Note: I initially found this behavior quite confusing but it turns out we perform this
pattern in the Authorize.net, WorldPay, and BraintreeBlue adapters. These changes follow
the same pattern.

Unit:
24 tests, 126 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote:
30 tests, 200 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed