Stripe: Add `exchange_rate` parameter (#2796)
This adds support for the optional `exchange_rate` parameter during
capture and purchase. This specifies an exchange rate to be used during
a transaction, and if incorrect or stale, Stripe will return a 400
`invalid_request_error` with an `invalid_exchange_rate` code. The
response also contains the updated rate in the event of a failure in the
`new_rate` parameter.

This is currently a preview feature.

Ref: https://stripe.com/docs/api#capture_charge

Unit:
126 tests, 667 assertions, 0 failures, 0 errors, 0 pendings, 0
omissions, 0 notifications
100% passed

Remote:
64 tests, 291 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
100% passed