SafeCharge: Additional gateway options
Adds merchant descriptor options that may be used to display on a
customer's credit card statement. Additionally, this adds support for
the new `sg_StoredCredentialMode` field. When `stored_credential_mode` is
true, this tells SafeCharge that the transaction was completed
using a previously tokenized and stored card data. When
`stored_credential_mode` is false (the default), this tells SafeCharge
the card was entered for the first time and will be stored for future
transactions.

Remote:
23 tests, 68 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Unit:
20 tests, 96 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed