QBMS: Support transcript scrubbing
We do not have test credentials for QBMS and they claim their test
sandbox is not currently working, so the remote tests cannot be run. The
mock transcripts added here in the unit tests should be accurate; they
come from an external scrub test that dates to the same time and API
version for the gateway implementation here in Active Merchant. This
scrub method is the same as what Spreedly calls manually on QBMS
transactions made today, and we've confirmed that those transcripts are
properly scrubbed.

Closes #2849

Unit:
15 tests, 74 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed