Fix Style/StringLiteral violations
Yes, I cheated and used rubocop to help with this, but I also did
actually audit the result, and it's pretty clean to read through with a
good diff tool since no spacing, etc. was changed. My recommendation
for review would be to look for and notice the absence of `#{}` in any
of the changed strings, meaning that the `"` to `'` conversion is safe.

Unit: 3874 tests, 67956 assertions, 0 failures, 0 errors, 0 pendings, 2 omissions, 0 notifications
100% passed