CyberSource: do not raise on HTML responses
Sometimes, CyberSource will 500, and return an HTML response body.
But that's a problem, because we assume all responses are XML. There
was a previous attempt to handle this situation in ba586a28b4d1, but
it won't work if an exception (e.g. a 400 or 500) is thrown as part
of the response, since the rescue block will be catching at the wrong
scope. Instead, capture the response or the exception, and then parse
that as a separate step, so we can be sure to catch the XML parse error
regardless.

Unit: 46 tests, 222 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote: didn't bother due to comment in fixtures.yml