BlueSnap: Handle 403 responses
403 Forbidden responses due to insufficient account permissions were
causing parsing exception errors. This handles them with a bespoke
response passing the raw text as message.

This also fixes the verify_credentials call; a change at the gateways
caused the response given for a bogus url (url/nonexistent) to be a
success with a different message. Hitting the root url now seems
sufficient to provoke a 401.

Closes #2948

Remote (1 unrelated failure):
25 tests, 72 assertions, 1 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
96% passed

Unit:
19 tests, 72 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed