Braintree: Add support for GooglePay
Adds support for processing GooglePay payment methods. It appears that
Braintree treats GooglePay in the same manner as AndroidPay. Even their
documentation (linked below) claims that GooglePay cards are represented
as AndroidPay cards. As such, the only way to get the GooglePay remote
test to pass is to pass the payment information in as `android_pay`.

Ref: https://developers.braintreepayments.com/guides/google-pay/server-side/ruby

Remote:
63 tests, 361 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Unit:
50 tests, 124 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed