Revert "Stripe: support shipping info in purchases"
This reverts commit d4d927e42a6ac7c0ce15be7e9a14f6c2f00a0ff5.
As-written, this could cause transaction to fail in some situations:
Stripe will validate *all* shipping parameters if *any* are present, so
attempting a purchase with some but not all shipping params could cause
the entire purchase to be rejected. The proper fix is to not include
any shipping details unless all are present, but for now, simply remove
the functionality entirely.