Braintree: Adds device_data
According to Braintree docs, `device_data` can be sent as a top-level
parameter when creating a `transaction`, `payment_method`, or
`customer`. And after chatting with their support, there is no way to
test this from the server-side. Deploying this with the hope that a
customer will give us some feedback about its efficacy.

Unit tests:
54 tests, 137 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
100% passed

Remote tests:
63 tests, 361 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
100% passed