Experiment: TreeCopy
This is an experiment with having a generic tree copier as a way
to simplify gateway implementations. It's not done, and may be
overengineered, but it's hopefully an interesting thought-experiment
on what a more declarative version of Active Merchant might look like.
The new TreeCopy module is largely extracted from what was left of me
trying to add this to Stripe on the last time I was in there.