Adyen: Add normalized hash of 3DS 2.0 data fields from web browsers
In an effort to begin supporting the intiation of 3DS 2.0 transactions,
this introduces another normalized hash for web browser data that can be
mapped to the fields a gateway is expecting. Similarly to stored
credentials, this hash should be used across all gateways in which 3DS
2.0 browser data is meant to be submitted so there is a single interface
for the client/device data.

Eventually I suspect this hash should also support mobile app information
by changing out the "channel" field and doing additional logic around that,
but for now we're starting with web browsers only.

The structure of the hash is as follows:

three_ds_2: {
  channel: "browser",
  browser_info: {
    accept_header: "unknown",
    depth: 100,
    java: false,
    language: "US",
    height: 1000,
    width: 500,
    timezone: "-120",
    user_agent: "unknown"
  }
}

This should be all of the data gateways are expecting, however as with
all payment gateways I imagine there are some that will accept or
require additional data. In those cases I suspect we can add that data
to this additional hash for a specific gateways needs.