Do not attempt application fee refund if refund was not successful (#3206)
This was a regression made in adc5a88c485e8de8cf30c064233fbe7bc927bf86
which was trying to make the App Fee refunds not impact the overall
status of the Refund response. That was achieved, but it also made the
unintentional change of always trying to refund the application fee
regardless of the refund succeeding or not. There was previously no test
coverage of this behaviour.