MiGS: Add tx_source gateway specific field
Added the `tx_source` gateway specific field to the MiGS gateway along
with remote and unit tests.

Based on feedback, the `add_tx_source` method wasn't added to the
`purchase_offsite_url` action since Spreedly is not in control of this
action and since there is not a way to run a remote test for it as the
`test_server_purchase_url` is the remote test that is still resulting
in an error.

No remote test was added for the `refund` action passing the `tx_source`
parameter because the main remote test for the `refund` action is
currently commented out and not working.

For consistency, all of the hash rockets in the unit tests were changed
to use the modern hash syntax, i.e., `key: value`.

CE-34 / CE-44

Unit:
9 tests, 21 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
100% passed

Remote:
16 tests, 59 assertions, 0 failures, 1 errors, 0 pendings, 0 omissions,
0 notifications
93.75% passed
Undefined method `request_uri` error in `test_server_purchase_url`;
error unrelated to change