Add Kosovo to the list of countries (#3226)
Although these codes are not officially in the ISO standard, they are
used by the European Commission, and we're seeing some buyers use them.
The numeric code, which is almost never used by the provider, is one of
the reserved codes that will never be assigned to an actual country. If
Kosovo is added to the ISO standard, we can correct it at that time.