Stripe Payment Intents - Add new gateway
In preparation for Stripe 3DS2 support, this PR adds support for Stripe
Payment Intents. Stripe Payment Intents is added as a gateway, with
functionality to create new PaymentIntents and Stripe PaymentMethods.
PaymentIntents can also be confirmed, captured, cancelled, shown, and
updated. This gateway is implemented as a subclass of the Stripe gateway.
To take advantage of the latest features, the API for this gateway is
bumped up to the latest version, `2019-05-16`.

Remote:
24 tests, 103 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Unit:
6 tests, 42 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Closes #3290