Cybersource: Allow partial update of credit card
When sending partial card details to Cybersource, an account number is
not required. The presence of the `accountNumber` element in the request
forces validation, and so a "Invalid account number. (231)" response
will be given when the account number is not provided, even though it is
optional for update.

This change ensures that the `<accountNumber/>` element is only included
in the request if there is a card number provided.