Decidir: Add Cabal card
Adds the Cabal card type to the Decidir gateway.

With our current test gateway implementation, it does not seem possible
to test Cabal cards. When attempting to run a remote test with a Cabal
test card number, the error message states that the request contains an
invalid BIN (first 6 digits of the card number). Therefore, the addition
of Cabal cards has not been tested on the Decidir gateway.

CE-94 / CE-100

Unit:
22 tests, 112 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
100% passed

Remote:
17 tests, 60 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
100% passed