Adyen: Send "NA" instead of "N/A"
When an address field is not provided, the default will now be "NA"
instead of the previous "N/A". According to Adyen, this should improve
acceptance rates for transactions that do not include an address
provided by the customer.

CE-115

Unit:
47 tests, 228 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
100% passed

Remote:
66 tests, 213 assertions, 2 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
96.9697% passed

The two tests that are failing:
 - test_successful_purchase_with_auth_data_via_threeds1_standalone
 - test_successful_purchase_with_auth_data_via_threeds2_standalone

These failures seem to be unrelated to the changes here.