Stripe Payment Intents: Successfully set application fee on capture
You do not need to pass the transfer destination in a capture request,
and in fact the request will fail if you try to post that data.
However, before this PR, we were requiring that a destination account
be present before setting the associated fee/transfer amount. This PR
updates the logic to set only the application fee or transfer amount
on capture requests when present.

Remote:
29 tests, 124 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Unit:
6 tests, 42 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed