Adyen: Add Cabal card
Added the Cabal card to the Adyen gateway, including adding unit and
remote tests.

Currently, Adyen does not support recurring transactions for Cabal
cards, so additional logic was added to the `store` method to return an
error message when attempting to store a Cabal card. Since the `store`
method for Adyen uses the `authorise` action, even though Cabal cards
cannot be successfully stored, the response from the gateway is
successful and does not return any sort of error message. Therefore, the
response from the gateway is altered to accurately reflect when a card
is not successfully stored by Adyen.

CE-99

Unit:
40 tests, 193 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
100% passed

Remote:
72 tests, 232 assertions, 3 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
95.8333% passed

The two failing tests are
 - `test_successful_purchase_with_auth_data_via_threeds1_standalone`
 - `test_successful_purchase_with_auth_data_via_threeds2_standalone`

These two tests were recently added in #3294 and appear to have never
passed for Spreedly.