Visanet Peru: Add amount argument to Capture
Capture was initially implemented without an amount argument, this adds
it for the sake of standardization.

Remote tests are causing 500s with our latest test credentials, but I'm
comfortable with only unit tests since this is a fix which doesn't have
any logic behind it (we're not actually adding the amount to the
request).

Unit:
15 tests, 75 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed