Stripe: Remove outdated 'customer options' deprecation
This is no longer deprecated; in fact it's actually required.

From the Stripe docs:

**source**: A payment source to be charged. This can be the ID of a card
(i.e., credit or debit card), a bank account, a source, a token, or a
connected account.

** For certain sources—namely, cards, bank accounts,
and attached sources—you must also pass the ID of the associated customer.**

https://stripe.com/docs/api/charges/create#create_charge-source

Closes #2421 (and removes a ton of of deprecation notices from log files
all across the world)

Unit:
136 tests, 728 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote:
67 tests, 312 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Closes #3401