Quickbooks: Add OAuth 2.0 support, support void
Quickbooks is deprecating OAuth 1.0 support, and requiring OAuth
2.0 no later than December 17, 2019. This PR adds support for
OAuth 2.0, requiring the access_token, refresh_token, client_id
and client_secret when creating a gateway. When refresh_token is
present, OAuth 2.0 authentication will be attempted. If not
included, OAuth 1.0 will be attempted and should still work until
the deprecation date.

In the Quickbooks implementation, the access_token expires every
60 minutes and you must use the refresh_token to generate a new
one. There's no way to proactively check if the token is still
valid; you'll just see that the request fails with an
`AuthenticationFailed` error. This PR adds logic to each action to
check for this error, and attempt to refresh the token and reissue
the request when it happens.

This gateway also did not previously implement the void method;
support is added here. It required a change to authorization_from
method to include the request-id since that is needed in the void
request. The authorization value is split appropriately to parse out
the value needed for a given method.

Remote:
14 tests, 37 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Unit:
21 tests, 114 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed