Worldpay: Update cardholder name logic based on 3DS version
In testing with Worldpay, there was some unexpected behavior when
the 3DS version was 2. The existing regex match didn't capture the
value of `2` as being equivalent to a 3DS2 version, and so the
cardholder name was getting set to '3D'. This update uses
`start_with?` to capture any version value that starts with `2`,
including `2`. Unit tests have also been updated to test various
permutations of 3DS2 version values.

Remote:
56 tests, 240 assertions, 2 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
96.4286% passed
(pre-existing failures)

Unit:
69 tests, 411 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed