Cardstream: Localize zero-decimal currencies
Zero-decimal currencies were being handled in an unexpected way on
Cardstream compared to other gateways. For instance, whereas we
expect a merchant to pass in an amount like 100 to be charged 1
JPY, the gateway processing handles this as though it's a charge
of 100 JPY. This PR uses localized_amount to divide the value by
100 before passing along to the gateway. The gateway does not
support any three-decimal currencies, so this PR does not try to
test those scenarios.

This gateway also had quite a few failing tests due to cc year,
updated error messages and some duplicate transaction handling.
These are also updated in this PR. The tests are still somewhat
temperamental in that you'll sometimes get a timeout, but that
does not appear to be related to these updates as it's a different
test or two each time (and sometimes none at all).

Remote:
31 tests, 204 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Unit:
26 tests, 132 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed