Remove customer hash when using a payment_method_nonce
When processing 3DS based requests there is a need to create a transaction
using a payment method nonce that is generated on the client. To facilitate that
need and keep existing behaviors with payment method nonces, a check for a
string based payment method nonce was added right before finalizing sale
parameters. If the payment method nonce is a string, the customer hash will be
removed since there should already be a customer in Braintree's system.

For further information regarding the need to use a client payment method nonce,
please check here: https://developers.braintreepayments.com/guides/3d-secure/server-side/ruby

Unit:

77 tests, 178 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote:

80 tests, 435 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed